feat: added perfomance sales

+++ owehub/owehub-main/services/apiGetAdderData.go
@@ -147,7 +147,7 @@ func HandleGetAdderDataRequest(resp http.ResponseWriter, req *http.Request) {
 			PerKwAmt = 0.0
 		}
 
-		RepPercent, ok := item["rep_percent"].(float64)
+		RepPercent, ok := item["rep_percent"].(int64)
 		if !ok {
 			log.FuncErrorTrace(0, "Failed to get RepPercent for Record ID %v. Item: %+v\n", RecordId, item)
 			RepPercent = 0.0
